create shared mock module for tests and fuzzer
